+++ src/plugin/cursor.js
@@ -13,6 +13,9 @@
  * @property {boolean} showTime=false Show the time on the cursor.
  * @property {object} customShowTimeStyle An object with custom styles which are
  * applied to the cursor time element.
+ * @property {string} showTimeYPosition='fixed' Use 'relative' to make the time on
+ * the cursor follow the x and the y-position of the mouse. Use 'fixed' to make the
+ * it only follow the x-position of the mouse.
  * @property {function} formatTimeCallback Formats the timestamp on the cursor.
  */
 
@@ -80,13 +83,26 @@ export default class CursorPlugin {
         customStyle: {},
         customShowTimeStyle: {},
         showTime: false,
+        showTimeYPosition: 'fixed',
         formatTimeCallback: null
     };
 
     /** @private */
     _onMousemove = e => {
         const bbox = this.wavesurfer.container.getBoundingClientRect();
-        this.updateCursorPosition(e.clientX - bbox.left);
+        let y = 0;
+        let x = e.clientX - bbox.left;
+
+        if (this.params.showTime) {
+            switch (this.params.showTimeYPosition) {
+                case 'relative':
+                    // follow y-position of the mouse
+                    y = e.clientY - (bbox.top + bbox.height / 2);
+                    break;
+            }
+        }
+
+        this.updateCursorPosition(x, y);
     };
     /** @private */
     _onMouseenter = () => this.showCursor();
@@ -215,11 +231,12 @@ export default class CursorPlugin {
     /**
      * Update the cursor position
      *
-     * @param {number} pos The x offset of the cursor in pixels
+     * @param {number} xpos The x offset of the cursor in pixels
+     * @param {number} ypos The y offset of the cursor in pixels
      */
-    updateCursorPosition(pos) {
+    updateCursorPosition(xpos, ypos) {
         this.style(this.cursor, {
-            left: `${pos}px`
+            left: `${xpos}px`
         });
         if (this.params.showTime) {
             const duration = this.wavesurfer.getDuration();
@@ -232,10 +249,11 @@ export default class CursorPlugin {
                 (duration / this.wavesurfer.drawer.width) * scrollWidth;
 
             const timeValue =
-                Math.max(0, (pos / elementWidth) * duration) + scrollTime;
+                Math.max(0, (xpos / elementWidth) * duration) + scrollTime;
             const formatValue = this.formatTime(timeValue);
             this.style(this.showTime, {
-                left: `${pos}px`
+                left: `${xpos}px`,
+                top: `${ypos}px`
             });
             this.displayTime.innerHTML = `${formatValue}`;
         }